+++ apps/platform/src/components/integrations/integrationTriggerList/index.tsx
@@ -1,6 +1,6 @@
 import type { Integration, IntegrationRun } from '@keyshade/schema'
-import { ErrorSVG, PendingSVG, VectorSVG } from '@public/svg/shared'
 import React, { useCallback } from 'react'
+import { ErrorSVG, PendingSVG, VectorSVG } from '@public/svg/shared'
 import ControllerInstance from '@/lib/controller-instance'
 import { formatDate, formatTime } from '@/lib/utils'
 import {